+++ src/test/java/cart/domain/product/ImageUrlTest.java
@@ -1,9 +1,8 @@
-package cart.domain;
+package cart.domain.product;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import cart.domain.product.ImageUrl;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
